Added management command for anonymizing submissions & only show wagtail form if enabled
